+++ package.json
@@ -7,6 +7,7 @@
     "mongodb": "^3.3.3",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
+    "react-fontawesome": "^1.7.1",
     "react-router-dom": "^5.0.1",
     "react-scripts": "^3.0.1",
     "react-table": "^6.10.0"
